Added an anchor to sections
Linked the header of each feature tile to an `id`, and then added an `id` attribute to the divider above it. This way, if you click on the header, it will add the `id` of the element to the window path, which will take you to that element.

Also, created a class with CSS rules to prevent the linked header from changing appearances.

Issue #25