Added scripts for scrolling animation
Created some checks so that the scrolling icon shown on the Home page will only display after 1 second of no scrolling. It will display for 2 seconds, and then disappear. If the user scrolls within these 2 seconds, it'll disappear.

This is accomplished by using some `setTimeout` and jQuery's `fadeIn` and `fadeOut` functions, as well as a few `if` checks.

Issue #42