Fixed the full-screen view for Safari browsers
Edited the `viewBox` of the SVG logo to accommodate for how Safari handles SVG clipping. This prevents the SVG from taking up too much unneeded vertical space in the layout.

Also, adjusted padding/margin/positioning so the text and logo will be centered in the middle of the containing `div`.

Issue #46