Changed the layout to center anchors in the middle of the screen
Changed the rows of the anchors to be the full height of the screen and removed the featurette divider, which was originally setting them apart with a margin. By doing this, we get each anchor section to be full screen on most screen sizes except phones, where the content will stretch the `div`.

Also, I minorly adjusted the size of the `vh` for the containers used at the top of full height containers.

Since we no longer needed the dividers, I removed them and attached the `id` to the `row` containing the anchor. Also, on mobile, I made sure to set a margin on the rows to keep them apart from each other a little bit.

Issue #58